// front/src/app/models/profile/statistics.model.ts
export class StatisticsModel {
    rank: number;
    gamesPvP: number;
    winGamesPvP: number;
    weekGamesPvP: number;
    weekWinGamesPvP: number;
    monthGamesPvP: number;
    monthWinGamesPvP: number;
    gamesPvE: number;
    winGamesPvE: number;
    weekGamesPvE: number;
    weekWinGamesPvE: number;
    monthGamesPvE: number;
    monthWinGamesPvE: number;
}